moving all functions into its own file and adding comments, menu has been moved into its own folder as well
